+++ project/src/measurement/MeasurementManager.java
@@ -5,6 +5,7 @@ import Exceptions.SerialCommunicationExceptions.PicaxeConnectionErrorException;
 import Exceptions.SpectrometerExceptions.SpectrometerNotConnected;
 import javafx.scene.control.Label;
 import javafx.scene.paint.Color;
+import serialCommunication.Spectrometer;
 import serialCommunication.StepperMotor;
 import com.oceanoptics.omnidriver.api.wrapper.Wrapper;
 import gui.chart.Chart;
@@ -20,20 +21,20 @@ import java.util.Set;
 public class MeasurementManager {
 
     private Timeline livemodeTimeline;
-    private Timeline seriesOfMTimeline;
     private int remainingSteps;
 
     public Wrapper wrapper;
-
+    private Spectrometer spectrometer;
     private StepperMotor stepperMotor;
 
-    public SpectroSimulator spectroSimulator;
+    //public SpectroSimulator spectroSimulator;
 
 
     public MeasurementManager(StepperMotor stepperMotor) {
         this.stepperMotor = stepperMotor;
         wrapper = new Wrapper();
-        spectroSimulator = new SpectroSimulator(200,800);
+        spectrometer = new Spectrometer(wrapper);
+        //spectroSimulator = new SpectroSimulator(200,800);
     }
 
     public void startLiveMode(Integer integrationTime, Chart chart){
@@ -49,6 +50,10 @@ public class MeasurementManager {
         livemodeTimeline.play();
     }
 
+    public void stopLiveMode(){
+        livemodeTimeline.stop();
+    }
+/*
     public void startSimulatedLiveMode(Integer integrationTime, Chart chart){
         spectroSimulator = new SpectroSimulator(200,800);
         Double interval = ((integrationTime/1000) + chart.getDrawingTime());
@@ -59,141 +64,19 @@ public class MeasurementManager {
         }));
         livemodeTimeline.setCycleCount(Timeline.INDEFINITE);
         livemodeTimeline.play();
-    }
+    }*/
 
-    public void stopLiveMode(){
-        livemodeTimeline.stop();
-    }
-
-    public void seriesOfMeasurements(Chart chart, Label currentAngleLabel, Label remainingStepsLabel) throws PicaxeConnectionErrorException, SpectrometerNotConnected {
-        if(stepperMotor.checkPicaxeConnection() == false){
-            throw new PicaxeConnectionErrorException("Picaxe is not connected");
-        }
-        checkConnectionOfSpectrometer();
-        //startSeriesOfMeasurements(chart, currentAngleLabel, remainingStepsLabel);
-        moveAndStartSeries(chart, currentAngleLabel, remainingStepsLabel);
-    }
-
-    private void moveAndStartSeries(Chart chart, Label currentAngleLabel, Label remainingStepsLabel){
-        double angle = Settings.getMeasurementMinAngle();
-        Timeline moving;
-        if (stepperMotor.currentAngle < angle) {
-            moving = new Timeline(new KeyFrame(Duration.millis(stepperMotor.getImpulseTime()), e -> {
-                stepperMotor.moveOnePulseForward(currentAngleLabel);
-            }));
-        } else {
-            moving = new Timeline(new KeyFrame(Duration.millis(stepperMotor.getImpulseTime()), e -> {
-                stepperMotor.moveOnePulseBackwards(currentAngleLabel);
-            }));
-        }
-        moving.setCycleCount(stepperMotor.pulsesNeededToMove(angle));
-        moving.setOnFinished(e -> startSeries(chart, currentAngleLabel, remainingStepsLabel));
-        moving.play();
-    }
-
-    private void startSeries(Chart chart, Label currentAngleLabel, Label remainingStepsLabel) {
-        Double interval = (Settings.getIntegrationTime()/1000) + chart.getDrawingTime() + stepperMotor.getStepTime();
-        Double startAngle = Settings.getMeasurementMinAngle();
-        Double endAngle = Settings.getMeasurementMaxAngle();
-
-        Integer stepsToDo = stepperMotor.stepsNeededToMove(endAngle);
-        remainingSteps = stepsToDo;
-        remainingStepsLabel.setText(String.valueOf(remainingSteps));
-
-        double[] wavelengths = wrapper.getWavelengths(0);
-        chart.setxValues(wavelengths);
-        wrapper.setIntegrationTime(0, Settings.getIntegrationTime());
-        wrapper.setScansToAverage(0, Settings.getNumberOfScansToAverage());
-
-        SeriesOfMeasurements seriesOfMeasurements = new SeriesOfMeasurements();
-        seriesOfMTimeline = new Timeline(new KeyFrame(Duration.millis(interval), e -> {
-
-            measureAndVisualize(chart, wavelengths, seriesOfMeasurements);
-
-            if (startAngle < endAngle){
-                stepperMotor.stepForward(currentAngleLabel);
-            } else {
-                stepperMotor.stepBackwards(currentAngleLabel);
-            }
-            remainingSteps --;
-            remainingStepsLabel.setText(String.valueOf(remainingSteps));
-        }));
-
-        seriesOfMTimeline.setCycleCount(stepsToDo);
-        seriesOfMTimeline.play();
-        seriesOfMTimeline.setOnFinished(e -> {
-            try {
-                measureAndVisualize(chart, wavelengths, seriesOfMeasurements); //odmeriam aj na konci intervalu
-                seriesOfMeasurements.save();
-                startLiveMode(Settings.getIntegrationTime(), chart);
-            } catch (ParameterIsNullException parameterIsNullException) {
-                parameterIsNullException.printStackTrace();
-            }
-        });
-    }
-
-    public void stopSeriesOfMeasurements(){
-        seriesOfMTimeline.stop();
-    }
-
-    private void measureAndVisualize(Chart chart, double[] wavelengths, SeriesOfMeasurements sofm){
-        double[] spectralData = wrapper.getSpectrum(0);
-        substractBackgroundIfNeeded(spectralData);
-        chart.replaceMainData(spectralData, "last measured data");
-        try {
-            sofm.addMeasurement(new Measurement(spectralData, wavelengths, stepperMotor.currentAngle));
-        } catch (ParameterIsNullException parameterIsNullException) {
-            parameterIsNullException.printStackTrace();
-        }
+    public void startSeriesOfMeasurements(Chart chart, Label currentAngleLabel, Label remainingStepsLabel) throws PicaxeConnectionErrorException, SpectrometerNotConnected {
+        SeriesOfMeasurements sofm = new SeriesOfMeasurements(wrapper, stepperMotor, spectrometer, this);
+        sofm.begin(chart, currentAngleLabel, remainingStepsLabel);
     }
 
     public void measureBackground(){
         Settings.setBackground(wrapper.getSpectrum(0));
     }
 
-    private void substractBackgroundIfNeeded(double[] values){
-        double[] background = Settings.getBackground();
-        if(Settings.getSubtractBackground() == false || background == null){
-            return;
-        }
-        for (int i = 0; i < Math.min(values.length, background.length); i++) {
-            values[i] = Math.max(values[i] - background[i], 0);
-        }
-    }
-
     public void checkConnectionOfSpectrometer() throws SpectrometerNotConnected {
-        if(wrapper == null){
-            throw new SpectrometerNotConnected("Spectrometer is not connected");
-        }
-
-        int numberOfSpectrometers;
-        try{
-            numberOfSpectrometers = wrapper.openAllSpectrometers();
-        } catch (java.lang.ExceptionInInitializerError | java.lang.NoClassDefFoundError e){
-            throw new SpectrometerNotConnected("Spectrometer is not connected");
-        }
-
-        if(numberOfSpectrometers == -1){ //nejaka specialna chyba
-            throw new SpectrometerNotConnected(wrapper.lastException.getMessage());
-        }
-        if(numberOfSpectrometers == 0){
-            throw new SpectrometerNotConnected("Spectrometer is not connected");
-        }
-        if(numberOfSpectrometers > 1){
-            throw new SpectrometerNotConnected("Multiple spectrometers are connected");
-        }
+        spectrometer.checkConnection();
     }
 
-//    public static void main(String[] args) {
-////        MeasurementManager mm = new MeasurementManager(serialCommManager);
-////        ChartManager chartManager = new ChartManager(mm.wrapper.getWavelengths(0),"wavelengths","intensities","Test");
-////        LineChart lineChart = chartManager.getComponent();
-////        VBox vbox = new VBox(lineChart);
-////
-////        Scene scene  = new Scene(vbox,1500,600);
-////        scene.getStylesheets().add("style.css");
-////        /*primaryStage.setTitle("chartTest"); 	// pomenuj okno aplikacie, javisko
-////        primaryStage.setScene(scene); 			// vloz scenu do hlavneho okna, na javisko
-////        primaryStage.show();*/
-//    }
 }
